+++ test/HAaveProtocolV2.test.js
@@ -37,6 +37,7 @@ const IAToken = artifacts.require('IATokenV2');
 const ILendingPool = artifacts.require('ILendingPoolV2');
 const IProvider = artifacts.require('ILendingPoolAddressesProviderV2');
 const SimpleToken = artifacts.require('SimpleToken');
+const ATOKEN_DUST = ether('0.00001');
 
 contract('Aave V2', function([_, user, someone]) {
   const aTokenAddress = ADAI_V2_TOKEN;
@@ -91,10 +92,10 @@ contract('Aave V2', function([_, user, someone]) {
           from: user,
           value: value,
         });
-        expect(await balanceProxy.get()).to.be.zero;
+        expect(await balanceProxy.get()).to.be.bignumber.zero;
         expect(
           await this.awmatic.balanceOf.call(this.proxy.address)
-        ).to.be.zero;
+        ).to.be.bignumber.zero;
         expectEqWithinBps(await this.awmatic.balanceOf.call(user), value, 100);
         expect(await balanceUser.delta()).to.be.bignumber.eq(
           ether('0')
@@ -113,10 +114,10 @@ contract('Aave V2', function([_, user, someone]) {
           from: user,
           value: value,
         });
-        expect(await balanceProxy.get()).to.be.zero;
+        expect(await balanceProxy.get()).to.be.bignumber.zero;
         expect(
           await this.awmatic.balanceOf.call(this.proxy.address)
-        ).to.be.zero;
+        ).to.be.bignumber.zero;
         expectEqWithinBps(await this.awmatic.balanceOf.call(user), value, 100);
         expect(await balanceUser.delta()).to.be.bignumber.eq(
           ether('0')
@@ -146,8 +147,10 @@ contract('Aave V2', function([_, user, someone]) {
           from: user,
           value: ether('0.1'),
         });
-        expect(await balanceProxy.get()).to.be.zero;
-        expect(await this.aToken.balanceOf.call(this.proxy.address)).to.be.zero;
+        expect(await balanceProxy.get()).to.be.bignumber.zero;
+        expect(
+          await this.aToken.balanceOf.call(this.proxy.address)
+        ).to.be.bignumber.zero;
         expectEqWithinBps(await this.aToken.balanceOf.call(user), value, 100);
         expect(await balanceUser.delta()).to.be.bignumber.eq(
           ether('0').sub(new BN(receipt.receipt.gasUsed))
@@ -173,8 +176,10 @@ contract('Aave V2', function([_, user, someone]) {
           from: user,
           value: ether('0.1'),
         });
-        expect(await balanceProxy.get()).to.be.zero;
-        expect(await this.aToken.balanceOf.call(this.proxy.address)).to.be.zero;
+        expect(await balanceProxy.get()).to.be.bignumber.zero;
+        expect(
+          await this.aToken.balanceOf.call(this.proxy.address)
+        ).to.be.bignumber.zero;
         expect(await this.aToken.balanceOf.call(user)).to.be.bignumber.eq(
           value
         );
@@ -246,7 +251,7 @@ contract('Aave V2', function([_, user, someone]) {
         // Verify proxy balance
         expect(
           await this.awmatic.balanceOf.call(this.proxy.address)
-        ).to.be.zero;
+        ).to.be.bignumber.zero;
         // Verify user balance
         // (deposit - withdraw) <= aTokenAfter < (deposit + interestMax - withdraw)
         expect(aTokenUserAfter).to.be.bignumber.gte(depositAmount.sub(value));
@@ -288,7 +293,7 @@ contract('Aave V2', function([_, user, someone]) {
         // Verify proxy balance
         expect(
           await this.awmatic.balanceOf.call(this.proxy.address)
-        ).to.be.zero;
+        ).to.be.bignumber.zero;
         // Verify user balance
         // (deposit - withdraw) <= aTokenAfter < (deposit + interestMax - withdraw)
         // NOTE: aTokenUserAfter == (depositAmount - withdraw - 1) (sometime, Ganache bug maybe)
@@ -348,8 +353,12 @@ contract('Aave V2', function([_, user, someone]) {
         // Verify handler return
         expect(value).to.be.bignumber.eq(handlerReturn);
         // Verify proxy balance
-        expect(await this.aToken.balanceOf.call(this.proxy.address)).to.be.zero;
-        expect(await this.token.balanceOf.call(this.proxy.address)).to.be.zero;
+        expect(
+          await this.aToken.balanceOf.call(this.proxy.address)
+        ).to.be.bignumber.zero;
+        expect(
+          await this.token.balanceOf.call(this.proxy.address)
+        ).to.be.bignumber.zero;
 
         // Verify user balance
         // (deposit - withdraw) <= aTokenAfter < (deposit + interestMax - withdraw)
@@ -396,8 +405,12 @@ contract('Aave V2', function([_, user, someone]) {
         expect(mulPercent(value, 101)).to.be.bignumber.gte(handlerReturn);
 
         // Verify proxy balance
-        expect(await this.aToken.balanceOf.call(this.proxy.address)).to.be.zero;
-        expect(await this.token.balanceOf.call(this.proxy.address)).to.be.zero;
+        expect(
+          await this.aToken.balanceOf.call(this.proxy.address)
+        ).to.be.bignumber.zero;
+        expect(
+          await this.token.balanceOf.call(this.proxy.address)
+        ).to.be.bignumber.zero;
         // Verify user balance
         // (deposit - withdraw -1) <= aTokenAfter < (deposit + interestMax - withdraw)
         // NOTE: aTokenUserAfter == (depositAmount - withdraw - 1) (sometime, Ganache bug maybe)
@@ -445,10 +458,14 @@ contract('Aave V2', function([_, user, someone]) {
         // Verify handler return
         expect(handlerReturn).to.be.bignumber.gte(depositAmount);
         // Verify proxy balance
-        expect(await this.aToken.balanceOf.call(this.proxy.address)).to.be.zero;
-        expect(await this.token.balanceOf.call(this.proxy.address)).to.be.zero;
+        expect(
+          await this.aToken.balanceOf.call(this.proxy.address)
+        ).to.be.bignumber.zero;
+        expect(
+          await this.token.balanceOf.call(this.proxy.address)
+        ).to.be.bignumber.zero;
         // Verify user balance
-        expect(aTokenUserAfter).to.be.zero;
+        expect(aTokenUserAfter).to.be.bignumber.lt(ATOKEN_DUST);
         expect(tokenUserAfter).to.be.bignumber.eq(handlerReturn);
         expect(await balanceUser.delta()).to.be.bignumber.eq(
           ether('0').sub(new BN(receipt.receipt.gasUsed))