+++ test/HAaveProtocolV2_flashloan.test.js
@@ -173,8 +173,8 @@ contract('AaveV2 flashloan', function([_, user, someone]) {
       });
 
       const fee = _getFlashloanFee(value);
-      expect(await balanceProxy.get()).to.be.zero;
-      expect(await this.tokenA.balanceOf.call(this.proxy.address)).to.be.zero;
+      expect(await balanceProxy.get()).to.be.bignumber.zero;
+      expect(await this.tokenA.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
       expect(await this.tokenA.balanceOf.call(user)).to.be.bignumber.eq(
         tokenAUser.add(value).sub(fee)
       );
@@ -214,8 +214,8 @@ contract('AaveV2 flashloan', function([_, user, someone]) {
         value: ether('0.1'),
       });
 
-      expect(await balanceProxy.get()).to.be.zero;
-      expect(await this.tokenA.balanceOf.call(this.proxy.address)).to.be.zero;
+      expect(await balanceProxy.get()).to.be.bignumber.zero;
+      expect(await this.tokenA.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
       expect(await this.tokenA.balanceOf.call(user)).to.be.bignumber.eq(
         tokenAUser.add(value).add(value)
       );
@@ -264,8 +264,8 @@ contract('AaveV2 flashloan', function([_, user, someone]) {
         value: ether('0.1'),
       });
 
-      expect(await balanceProxy.get()).to.be.zero;
-      expect(await this.tokenA.balanceOf.call(this.proxy.address)).to.be.zero;
+      expect(await balanceProxy.get()).to.be.bignumber.zero;
+      expect(await this.tokenA.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
       expect(await this.tokenA.balanceOf.call(user)).to.be.bignumber.eq(
         tokenAUser.add(value).add(value)
       );
@@ -300,9 +300,9 @@ contract('AaveV2 flashloan', function([_, user, someone]) {
         value: ether('0.1'),
       });
 
-      expect(await balanceProxy.get()).to.be.zero;
-      expect(await this.tokenA.balanceOf.call(this.proxy.address)).to.be.zero;
-      expect(await this.tokenB.balanceOf.call(this.proxy.address)).to.be.zero;
+      expect(await balanceProxy.get()).to.be.bignumber.zero;
+      expect(await this.tokenA.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
+      expect(await this.tokenB.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
 
       const fee = _getFlashloanFee(value);
       expect(await this.tokenA.balanceOf.call(user)).to.be.bignumber.eq(
@@ -509,9 +509,9 @@ contract('AaveV2 flashloan', function([_, user, someone]) {
         value: ether('0.1'),
       });
 
-      expect(await balanceProxy.get()).to.be.zero;
-      expect(await this.tokenA.balanceOf.call(this.proxy.address)).to.be.zero;
-      expect(await this.tokenB.balanceOf.call(this.proxy.address)).to.be.zero;
+      expect(await balanceProxy.get()).to.be.bignumber.zero;
+      expect(await this.tokenA.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
+      expect(await this.tokenB.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
 
       const fee = value
         .mul(new BN('9'))
@@ -572,9 +572,9 @@ contract('AaveV2 flashloan', function([_, user, someone]) {
         value: ether('0.1'),
       });
 
-      expect(await balanceProxy.get()).to.be.zero;
-      expect(await this.tokenA.balanceOf.call(this.proxy.address)).to.be.zero;
-      expect(await this.tokenB.balanceOf.call(this.proxy.address)).to.be.zero;
+      expect(await balanceProxy.get()).to.be.bignumber.zero;
+      expect(await this.tokenA.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
+      expect(await this.tokenB.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
 
       const fee = value
         .mul(new BN('9'))
@@ -650,9 +650,9 @@ contract('AaveV2 flashloan', function([_, user, someone]) {
         value: ether('0.1'),
       });
 
-      expect(await balanceProxy.get()).to.be.zero;
-      expect(await this.tokenA.balanceOf.call(this.proxy.address)).to.be.zero;
-      expect(await this.tokenB.balanceOf.call(this.proxy.address)).to.be.zero;
+      expect(await balanceProxy.get()).to.be.bignumber.zero;
+      expect(await this.tokenA.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
+      expect(await this.tokenB.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
 
       const fee = _getFlashloanFee(value);
       expect(await this.tokenA.balanceOf.call(user)).to.be.bignumber.eq(