+++ test/HCurveCrypto.test.js
@@ -89,10 +89,10 @@ contract('Curve Crypto', function([_, user]) {
         expect(handlerReturn).to.be.bignumber.lte(mulPercent(answer, 101));
 
         // Check proxy
-        expect(await balanceProxy.get()).to.be.zero;
-        expect(await token0.balanceOf.call(this.proxy.address)).to.be.zero;
-        expect(await token1.balanceOf.call(this.proxy.address)).to.be.zero;
-        expect(await token2.balanceOf.call(this.proxy.address)).to.be.zero;
+        expect(await balanceProxy.get()).to.be.bignumber.zero;
+        expect(await token0.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
+        expect(await token1.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
+        expect(await token2.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
 
         profileGas(receipt);
       });
@@ -206,11 +206,11 @@ contract('Curve Crypto', function([_, user]) {
         expect(handlerReturn).to.be.bignumber.lte(mulPercent(answer, 101));
 
         // Check proxy
-        expect(await balanceProxy.get()).to.be.zero;
-        expect(await token0.balanceOf.call(this.proxy.address)).to.be.zero;
-        expect(await token1.balanceOf.call(this.proxy.address)).to.be.zero;
-        expect(await token2.balanceOf.call(this.proxy.address)).to.be.zero;
-        expect(await poolToken.balanceOf.call(this.proxy.address)).to.be.zero;
+        expect(await balanceProxy.get()).to.be.bignumber.zero;
+        expect(await token0.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
+        expect(await token1.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
+        expect(await token2.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
+        expect(await poolToken.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
 
         profileGas(receipt);
       });